Big refactoring of regular skillcheck code
After the raylib update inexplicably broke everything, I tried debugging
but couldn't figure why everything broke. There was some really awful
code that needed refactoring anyway, so I did so, and the regular
skillcheck is working again. The others currently remain broken, fixing
them is a TODO.

In this refactor, I copied values for the width of the zones, and
spawning locations from the DBD wiki.